create views file and move routes
